+++ tests/test_module.rs
@@ -4,26 +4,33 @@
 extern crate pyo3;
 
 use pyo3::prelude::*;
+
+#[cfg(Py_3)]
 use pyo3::types::PyDict;
 
+#[cfg(Py_3)]
 #[macro_use]
 mod common;
 
 #[pyclass]
+#[cfg(Py_3)]
 struct EmptyClass {}
 
+#[cfg(Py_3)]
 fn sum_as_string(a: i64, b: i64) -> String {
     format!("{}", a + b).to_string()
 }
 
 #[pyfunction]
+#[cfg(Py_3)]
 /// Doubles the given value
 fn double(x: usize) -> usize {
     x * 2
 }
 
 /// This module is implemented in Rust.
-#[pymodinit]
+#[pymodule]
+#[cfg(Py_3)]
 fn module_with_functions(py: Python, m: &PyModule) -> PyResult<()> {
     #[pyfn(m, "sum_as_string")]
     fn sum_as_string_py(_py: Python, a: i64, b: i64) -> PyResult<String> {
@@ -47,6 +54,7 @@ fn module_with_functions(py: Python, m: &PyModule) -> PyResult<()> {
 }
 
 #[test]
+#[cfg(Py_3)]
 fn test_module_with_functions() {
     let gil = Python::acquire_gil();
     let py = gil.python();
@@ -71,7 +79,7 @@ fn test_module_with_functions() {
     run("assert module_with_functions.also_double.__doc__ == 'Doubles the given value'");
 }
 
-#[pymodinit(other_name)]
+#[pymodule(other_name)]
 fn some_name(_: Python, _: &PyModule) -> PyResult<()> {
     Ok(())
 }
@@ -125,16 +133,19 @@ fn test_module_from_code() {
 }
 
 #[pyfunction]
+#[cfg(Py_3)]
 fn r#move() -> usize {
     42
 }
 
-#[pymodinit]
+#[pymodule]
+#[cfg(Py_3)]
 fn raw_ident_module(_py: Python, module: &PyModule) -> PyResult<()> {
     module.add_wrapped(wrap_function!(r#move))
 }
 
 #[test]
+#[cfg(Py_3)]
 fn test_raw_idents() {
     let gil = Python::acquire_gil();
     let py = gil.python();
@@ -150,21 +161,21 @@ fn subfunction() -> String {
     "Subfunction".to_string()
 }
 
-#[pymodinit]
 #[cfg(Py_3)]
+#[pymodule]
 fn submodule(_py: Python, module: &PyModule) -> PyResult<()> {
     module.add_wrapped(wrap_function!(subfunction))?;
     Ok(())
 }
 
-#[pyfunction]
 #[cfg(Py_3)]
+#[pyfunction]
 fn superfunction() -> String {
     "Superfunction".to_string()
 }
 
-#[pymodinit]
 #[cfg(Py_3)]
+#[pymodule]
 fn supermodule(_py: Python, module: &PyModule) -> PyResult<()> {
     module.add_wrapped(wrap_function!(superfunction))?;
     module.add_wrapped(wrap_module!(submodule))?;